feat: Support sse to sse
