+++ internal/core/mcpproxy/common.go
@@ -1,0 +1,95 @@
+package mcpproxy
+
+import (
+	"encoding/json"
+
+	mcpgo "github.com/mark3labs/mcp-go/mcp"
+	"github.com/mcp-ecosystem/mcp-gateway/pkg/mcp"
+)
+
+// convertMCPGoResult converts a result from mcp-go format to local mcp format
+func convertMCPGoResult(mcpgoResult *mcpgo.CallToolResult) *mcp.CallToolResult {
+	result := &mcp.CallToolResult{
+		IsError: mcpgoResult.IsError,
+	}
+
+	// Process content items
+	if len(mcpgoResult.Content) > 0 {
+		var validContents []mcp.Content
+
+		for _, content := range mcpgoResult.Content {
+			// Skip null content
+			if content == nil {
+				continue
+			}
+
+			// Try to get content type
+			contentType := ""
+			switch c := content.(type) {
+			case *mcpgo.TextContent:
+				contentType = "text"
+				validContents = append(validContents, &mcp.TextContent{
+					Type: "text",
+					Text: c.Text,
+				})
+			case *mcpgo.ImageContent:
+				contentType = "image"
+				validContents = append(validContents, &mcp.ImageContent{
+					Type:     "image",
+					Data:     c.Data,
+					MimeType: c.MIMEType,
+				})
+			case *mcpgo.AudioContent:
+				contentType = "audio"
+				validContents = append(validContents, &mcp.AudioContent{
+					Type:     "audio",
+					Data:     c.Data,
+					MimeType: c.MIMEType,
+				})
+			default:
+				// Try to parse from raw content
+				rawContent, err := json.Marshal(content)
+				if err == nil {
+					var contentMap map[string]interface{}
+					if json.Unmarshal(rawContent, &contentMap) == nil {
+						if typ, ok := contentMap["type"].(string); ok {
+							contentType = typ
+
+							switch contentType {
+							case "text":
+								if text, ok := contentMap["text"].(string); ok {
+									validContents = append(validContents, &mcp.TextContent{
+										Type: "text",
+										Text: text,
+									})
+								}
+							case "image":
+								data, _ := contentMap["data"].(string)
+								mimeType, _ := contentMap["mimeType"].(string)
+								validContents = append(validContents, &mcp.ImageContent{
+									Type:     "image",
+									Data:     data,
+									MimeType: mimeType,
+								})
+							case "audio":
+								data, _ := contentMap["data"].(string)
+								mimeType, _ := contentMap["mimeType"].(string)
+								validContents = append(validContents, &mcp.AudioContent{
+									Type:     "audio",
+									Data:     data,
+									MimeType: mimeType,
+								})
+							}
+						}
+					}
+				}
+			}
+		}
+
+		if len(validContents) > 0 {
+			result.Content = validContents
+		}
+	}
+
+	return result
+}
