+++ internal/template/renderer.go
@@ -4,7 +4,9 @@ import (
 	"bytes"
 	"crypto/sha256"
 	"encoding/hex"
+	"encoding/json"
 	"fmt"
+	"net/http"
 	"text/template"
 )
 
@@ -49,4 +51,60 @@ func (r *Renderer) Render(tmpl string, ctx *Context) (string, error) {
 	}
 
 	return buf.String(), nil
+}
+
+// RenderTemplate renders a template with the given context
+func RenderTemplate(tmpl string, ctx *Context) (string, error) {
+	renderer := NewRenderer()
+	return renderer.Render(tmpl, ctx)
+}
+
+// PrepareTemplateContext prepares the template context with request and config data
+func PrepareTemplateContext(args map[string]any, request *http.Request, serverCfg map[string]string) (*Context, error) {
+	tmplCtx := NewContext()
+	tmplCtx.Args = preprocessArgs(args)
+
+	// Process request headers
+	for k, v := range request.Header {
+		if len(v) > 0 {
+			tmplCtx.Request.Headers[k] = v[0]
+		}
+	}
+
+	// Only process server config templates if serverCfg is provided
+	if serverCfg != nil {
+		// Process server config templates
+		for k, v := range serverCfg {
+			rendered, err := RenderTemplate(v, tmplCtx)
+			if err != nil {
+				return nil, fmt.Errorf("failed to render config template: %w", err)
+			}
+			serverCfg[k] = rendered
+		}
+		tmplCtx.Config = serverCfg
+	}
+
+	return tmplCtx, nil
+}
+
+func preprocessArgs(args map[string]any) map[string]any {
+	processed := make(map[string]any)
+
+	for k, v := range args {
+		switch val := v.(type) {
+		case []any:
+			ss, _ := json.Marshal(val)
+			processed[k] = string(ss)
+		case float64:
+			// If the float64 equals its integer conversion, it's an integer
+			if val == float64(int64(val)) {
+				processed[k] = int64(val)
+			} else {
+				processed[k] = val
+			}
+		default:
+			processed[k] = v
+		}
+	}
+	return processed
 }
